Move files into the abstraction project
