+++ src/NexusMods.Abstractions/CLI/IOptionParser.cs
@@ -1,4 +1,4 @@
-namespace NexusMods.CLI.OptionParsers;
+namespace NexusMods.Abstractions.CLI;
 
 /// <summary>
 /// A service that parses a string into a C# type for use with CLI inputs
@@ -13,7 +13,7 @@ public interface IOptionParser<T>
     /// <param name="definition"></param>
     /// <returns></returns>
     public T Parse(string input, OptionDefinition<T> definition);
-    
+
     /// <summary>
     /// Get various options that match the input string, used for tab completion
     /// </summary>
