+++ src/NexusMods.CLI/OptionDefinitionExtensions.cs
@@ -1,0 +1,7 @@
+﻿namespace NexusMods.CLI;
+
+public static class OptionDefinitionExtensions
+{
+
+
+}
